NOTICE_BOARD_APP/requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
celery==5.3.4
kombu==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.6.0
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0